+++ tests/integration/verify/v2/service/test_verification.py
@@ -43,22 +43,14 @@ class VerificationTestCase(IntegrationTestCase):
                 "valid": false,
                 "date_created": "2015-07-30T20:00:00Z",
                 "date_updated": "2015-07-30T20:00:00Z",
-                "lookup": {
-                    "carrier": {
-                        "error_code": null,
-                        "name": "Carrier Name",
-                        "mobile_country_code": "310",
-                        "mobile_network_code": "150",
-                        "type": "mobile"
-                    }
-                },
+                "lookup": {},
                 "amount": null,
                 "payee": null,
                 "send_code_attempts": [
                     {
                         "time": "2015-07-30T20:00:00Z",
                         "channel": "SMS",
-                        "channel_id": null
+                        "attempt_sid": "VLaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                     }
                 ],
                 "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Verifications/VEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
@@ -85,22 +77,14 @@ class VerificationTestCase(IntegrationTestCase):
                 "valid": false,
                 "date_created": "2015-07-30T20:00:00Z",
                 "date_updated": "2015-07-30T20:00:00Z",
-                "lookup": {
-                    "carrier": {
-                        "error_code": null,
-                        "name": "Carrier Name",
-                        "mobile_country_code": "310",
-                        "mobile_network_code": "150",
-                        "type": "mobile"
-                    }
-                },
+                "lookup": {},
                 "amount": null,
                 "payee": null,
                 "send_code_attempts": [
                     {
                         "time": "2015-07-30T20:00:00Z",
                         "channel": "whatsapp",
-                        "channel_id": null
+                        "attempt_sid": "VLaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                     }
                 ],
                 "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Verifications/VEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
@@ -127,22 +111,14 @@ class VerificationTestCase(IntegrationTestCase):
                 "valid": false,
                 "date_created": "2015-07-30T20:00:00Z",
                 "date_updated": "2015-07-30T20:00:00Z",
-                "lookup": {
-                    "carrier": {
-                        "error_code": null,
-                        "name": null,
-                        "mobile_country_code": null,
-                        "mobile_network_code": null,
-                        "type": null
-                    }
-                },
+                "lookup": {},
                 "amount": null,
                 "payee": null,
                 "send_code_attempts": [
                     {
                         "time": "2015-07-30T20:00:00Z",
                         "channel": "EMAIL",
-                        "channel_id": null
+                        "attempt_sid": "VLaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                     }
                 ],
                 "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Verifications/VEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
@@ -169,22 +145,14 @@ class VerificationTestCase(IntegrationTestCase):
                 "valid": false,
                 "date_created": "2015-07-30T20:00:00Z",
                 "date_updated": "2015-07-30T20:00:00Z",
-                "lookup": {
-                    "carrier": {
-                        "error_code": null,
-                        "name": "Carrier Name",
-                        "mobile_country_code": "310",
-                        "mobile_network_code": "150",
-                        "type": "mobile"
-                    }
-                },
+                "lookup": {},
                 "amount": null,
                 "payee": null,
                 "send_code_attempts": [
                     {
                         "time": "2015-07-30T20:00:00Z",
                         "channel": "SMS",
-                        "channel_id": null
+                        "attempt_sid": "VLaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                     }
                 ],
                 "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Verifications/VEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
@@ -226,22 +194,14 @@ class VerificationTestCase(IntegrationTestCase):
                 "valid": false,
                 "date_created": "2015-07-30T20:00:00Z",
                 "date_updated": "2015-07-30T20:00:00Z",
-                "lookup": {
-                    "carrier": {
-                        "error_code": null,
-                        "name": "Carrier Name",
-                        "mobile_country_code": "310",
-                        "mobile_network_code": "150",
-                        "type": "mobile"
-                    }
-                },
+                "lookup": {},
                 "amount": null,
                 "payee": null,
                 "send_code_attempts": [
                     {
                         "time": "2015-07-30T20:00:00Z",
                         "channel": "SMS",
-                        "channel_id": null
+                        "attempt_sid": "VLaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                     }
                 ],
                 "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Verifications/VEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
@@ -268,22 +228,14 @@ class VerificationTestCase(IntegrationTestCase):
                 "valid": true,
                 "date_created": "2015-07-30T20:00:00Z",
                 "date_updated": "2015-07-30T20:00:00Z",
-                "lookup": {
-                    "carrier": {
-                        "error_code": null,
-                        "name": "Carrier Name",
-                        "mobile_country_code": "310",
-                        "mobile_network_code": "150",
-                        "type": "mobile"
-                    }
-                },
+                "lookup": {},
                 "amount": null,
                 "payee": null,
                 "send_code_attempts": [
                     {
                         "time": "2015-07-30T20:00:00Z",
                         "channel": "SMS",
-                        "channel_id": null
+                        "attempt_sid": "VLaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                     }
                 ],
                 "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Verifications/VEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
@@ -322,22 +274,14 @@ class VerificationTestCase(IntegrationTestCase):
                 "valid": false,
                 "date_created": "2015-07-30T20:00:00Z",
                 "date_updated": "2015-07-30T20:00:00Z",
-                "lookup": {
-                    "carrier": {
-                        "error_code": null,
-                        "name": "Carrier Name",
-                        "mobile_country_code": "310",
-                        "mobile_network_code": "150",
-                        "type": "mobile"
-                    }
-                },
+                "lookup": {},
                 "amount": null,
                 "payee": null,
                 "send_code_attempts": [
                     {
                         "time": "2015-07-30T20:00:00Z",
                         "channel": "SMS",
-                        "channel_id": null
+                        "attempt_sid": "VLaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                     }
                 ],
                 "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Verifications/VEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"