+++ twilio/rest/taskrouter/v1/workspace/workspace_cumulative_statistics.py
@@ -137,7 +137,7 @@ class WorkspaceCumulativeStatisticsContext(InstanceContext):
         :param unicode minutes: Only calculate statistics since this many minutes in the past
         :param datetime start_date: Only calculate statistics from on or after this date
         :param unicode task_channel: Only calculate cumulative statistics on this TaskChannel
-        :param unicode split_by_wait_time: A comma separated list of values that describes the thresholds to calculate statistics on
+        :param unicode split_by_wait_time: A comma separated list of values that describes the thresholds, in seconds, to calculate statistics on
 
         :returns: The fetched WorkspaceCumulativeStatisticsInstance
         :rtype: twilio.rest.taskrouter.v1.workspace.workspace_cumulative_statistics.WorkspaceCumulativeStatisticsInstance
@@ -403,7 +403,7 @@ class WorkspaceCumulativeStatisticsInstance(InstanceResource):
         :param unicode minutes: Only calculate statistics since this many minutes in the past
         :param datetime start_date: Only calculate statistics from on or after this date
         :param unicode task_channel: Only calculate cumulative statistics on this TaskChannel
-        :param unicode split_by_wait_time: A comma separated list of values that describes the thresholds to calculate statistics on
+        :param unicode split_by_wait_time: A comma separated list of values that describes the thresholds, in seconds, to calculate statistics on
 
         :returns: The fetched WorkspaceCumulativeStatisticsInstance
         :rtype: twilio.rest.taskrouter.v1.workspace.workspace_cumulative_statistics.WorkspaceCumulativeStatisticsInstance